blog_platform/backend/requirements.txt
fastapi==0.85.1
uvicorn[standard]==0.13.4

# ORM & DB
tortoise-orm==0.17.8
aiomysql==0.2.0

# 认证与安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# HTTP 客户端
httpx==0.27.2

# 消息队列
aio-pika==9.5.6

# 配置与环境
python-dotenv==1.2.1
pydantic==1.10.24

# 日志
loguru==0.7.2